[REFACTOR] modularised & added deletesnackbar in gallery
